requirements.txt
aiogram
chromadb
openai
sentence-transformers
python-docx
pymupdf
python-dotenv
